Update sonos library, hackily update sonosPlayURI in sonos.js
