+++ lib/routes/spotify.js
@@ -8,19 +8,14 @@ const router = express.Router();
 
 // API
 
-import { sonosFetchActiveSonosMiddleware, sonosPlayURI, sonosCurrentState } from '../api/sonos';
+import { sonosPlayURI } from '../api/sonos';
 import { spotifySearch } from '../api/spotify';
 
 // Routes
 
-router.post('/play/:trackId', sonosFetchActiveSonosMiddleware, function (req, res) {
-
-	sonosPlayURI(req.params.trackId).then(function (json) {
-
-		res.json(json);
-
-	});
-
+router.post('/play/:trackId', (req, res) => {
+	sonosPlayURI(req.params.trackId)
+  res.json({success: true});
 });
 
 router.get('/search/:searchQuery', function (req, res) {
