+++ package.json
@@ -20,7 +20,7 @@
     "debug": "~2.2.0",
     "express": "~4.13.1",
     "q": "^1.5.0",
-    "sonos": "^0.14.1"
+    "sonos": "^1.5.1"
   },
   "devDependencies": {
     "babel-cli": "^6.14.0",
